requirements.txt
Django==5.2
PyMuPDF
openai
python-docx
python-dotenv
